Fix all /W4 warnings in tests and library headers
C4267 (cluster.h): Add static_cast<int> for size_t->int in centroid
  add/remove calls and the size() return value.

C4101 (test_gaussian/test_gamma): Remove named 'e' from catch clauses
  where the variable was only used inside assert() — a no-op in Release.

C4068/C4996 (test_optimized_matrix3d): Guard GCC diagnostic pragmas
  with #ifndef _MSC_VER; add matching #pragma warning(push/disable:4996)
  for the deprecated-method compatibility test.

C4244 (14 test files): Add static_cast<double> on size_t->Observation
  assignments in ObservationSet loops across training, calculator,
  integration, and linear-ops test files.

C4834 (4 test files): Replace EXPECT_NO_THROW calls that discarded
  [[nodiscard]] return values with substantive assertions (EXPECT_GE,
  EXPECT_EQ, EXPECT_FALSE) that also improve test coverage.

C4189 (15 distribution test files): Add #pragma warning(disable:4189)
  guarded by #ifdef _MSC_VER. These legacy assert()-based test files
  compute variables used only in debug assert() calls; the warning is
  correct but unavoidable without restructuring the tests.

All 36/36 tests pass. Zero warnings in Release build with /W4.

+++ tests/calculators/test_calculator_edge_cases.cpp
@@ -61,7 +61,7 @@ protected:
 TEST_F(CalculatorEdgeCasesTest, SingleState_FB_Finite) {
     auto hmm = make_single_state_hmm();
     ObservationSet obs(5);
-    for (std::size_t i = 0; i < 5; ++i) obs(i) = i % 4;
+    for (std::size_t i = 0; i < 5; ++i) obs(i) = static_cast<double>(i % 4);
 
     ForwardBackwardCalculator fbc(*hmm, obs);
     EXPECT_TRUE(std::isfinite(fbc.getLogProbability()));
@@ -71,7 +71,7 @@ TEST_F(CalculatorEdgeCasesTest, SingleState_FB_Finite) {
 TEST_F(CalculatorEdgeCasesTest, SingleState_Viterbi_OnlyState0) {
     auto hmm = make_single_state_hmm();
     ObservationSet obs(5);
-    for (std::size_t i = 0; i < 5; ++i) obs(i) = i % 4;
+    for (std::size_t i = 0; i < 5; ++i) obs(i) = static_cast<double>(i % 4);
 
     ViterbiCalculator vc(*hmm, obs);
     const auto seq = vc.decode();
@@ -154,7 +154,7 @@ TEST_F(CalculatorEdgeCasesTest, DeterministicEmissions_FB_NoNaN) {
 TEST_F(CalculatorEdgeCasesTest, LongSequence_FB_NoUnderflow) {
     auto hmm = make_single_state_hmm();
     ObservationSet longObs(500);
-    for (std::size_t i = 0; i < 500; ++i) longObs(i) = i % 4;
+    for (std::size_t i = 0; i < 500; ++i) longObs(i) = static_cast<double>(i % 4);
 
     ForwardBackwardCalculator fbc(*hmm, longObs);
     EXPECT_TRUE(std::isfinite(fbc.getLogProbability()));
@@ -167,7 +167,7 @@ TEST_F(CalculatorEdgeCasesTest, LongSequence_FB_NoUnderflow) {
 TEST_F(CalculatorEdgeCasesTest, LongSequence_Viterbi_NoErrors) {
     auto hmm = make_single_state_hmm();
     ObservationSet longObs(500);
-    for (std::size_t i = 0; i < 500; ++i) longObs(i) = i % 4;
+    for (std::size_t i = 0; i < 500; ++i) longObs(i) = static_cast<double>(i % 4);
 
     EXPECT_NO_THROW({
         ViterbiCalculator vc(*hmm, longObs);
